Fix new warning from Clang.

+++ src/Fonts/GlyphRendererOutline.cpp
@@ -98,7 +98,7 @@ GlyphRendererOutline::render(FT_ULong charcode)
               BLEND);
     
     const glm::ivec2 bearing(bitmapGlyphBorder->left, bitmapGlyphBorder->top);
-    const unsigned advance = getFace()->glyph->advance.x;
+    const long advance = getFace()->glyph->advance.x;
     
     FT_Done_Glyph(glyphInterior);
     FT_Done_Glyph(glyphBorder);
